% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{add_shortest_path}
\alias{add_shortest_path}
\title{function to add in the shortest path to the drugs_biomarkers_targets data frame}
\usage{
add_shortest_path(
  drugs_biomarkers_targets,
  inputs,
  inputs_in_path,
  cancer_path
)
}
\description{
function to add in the shortest path to the drugs_biomarkers_targets data frame
}
