% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{calculate_shortest_path}
\alias{calculate_shortest_path}
\title{function that calculates the shortest paths from some inputs to some outputs}
\usage{
calculate_shortest_path(inputs, outputs, igraph_obj)
}
\arguments{
\item{igraph_obj}{is created by converting a data frame of edges using graph.data.frame}

\item{input}{character vectors}

\item{output}{character vector}
}
\value{
a matrix with the inputs as rows, the outputs as columns, and the shortest path length as entries
}
\description{
function that calculates the shortest paths from some inputs to some outputs
}
