% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{calculate_shortest_path_from_inputs}
\alias{calculate_shortest_path_from_inputs}
\title{function that calculates the shortest paths from inputs to everything else that
is in a pathway but not in MP
calls on calculate_shortest_path}
\usage{
calculate_shortest_path_from_inputs(cancer_path, MP, inputs_in_path)
}
\description{
function that calculates the shortest paths from inputs to everything else that
is in a pathway but not in MP
calls on calculate_shortest_path
}
