% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-vis.R
\docType{class}
\name{CdgnetViz-class}
\alias{CdgnetViz-class}
\alias{CdgnetViz}
\title{Web component and container for the sankey visualization (cdgnet-viz).}
\description{
Web component and container for the sankey visualization (cdgnet-viz).
}
\section{Methods}{

\describe{
\item{\code{get_attributes()}}{Get attributes for rendering component}

\item{\code{get_colors()}}{Get chart colors}

\item{\code{get_component_type()}}{Get component type for prefix of random id generator}

\item{\code{get_data()}}{Get chart data}

\item{\code{get_default_colors()}}{Get default colors}

\item{\code{get_default_settings()}}{Get default component settings}

\item{\code{get_name()}}{html tag of visualization component}

\item{\code{get_settings()}}{Get chart settings}

\item{\code{render_component(shiny = FALSE)}}{Render component to html}

\item{\code{set_colors(colors)}}{Set chart colors}

\item{\code{set_data(data)}}{Set chart data}

\item{\code{set_settings(settings)}}{Modify current settings
\describe{
  \item{settings}{List of new settings.
    Call get_available_settings for settings available to modify.
  }
}}
}}

