% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{combine_drugs}
\alias{combine_drugs}
\title{function that combines all drugs from all categories}
\usage{
combine_drugs(Cat1, Cat2, Cat3, Cat4 = NULL)
}
\arguments{
\item{Cat1}{categeory 1 data.frame}

\item{Cat2, }{categeory 2 data.frame}

\item{Cat3, }{categeory 3 data.frame}

\item{Cat4}{categeory 4 data.frame, can be null}
}
\value{
merged data frame
}
\description{
function that combines all drugs from all categories
}
