% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{dataParser}
\alias{dataParser}
\title{Parse the result category into the ui json object}
\usage{
dataParser(input, fda, fda_other, rec)
}
\arguments{
\item{input}{molecular profile}

\item{fda}{fda approved drugs (cat 1)}

\item{fda_other}{fda approved drugs for other cancer types ( cat 2)}

\item{rec}{drug recommendations}
}
\value{
json object
}
\description{
Parse the result category into the ui json object
}
