% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-shiny.R
\docType{data}
\name{.cdgnetUI}
\alias{.cdgnetUI}
\title{Defines shiny user interface for the cdgnet applicatio}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
.cdgnetUI
}
\description{
Defines shiny user interface for the cdgnet applicatio
}
\keyword{datasets}
