% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DrugBank_targets}
\alias{DrugBank_targets}
\title{DrugBank_targets}
\format{
An object of class \code{data.frame} with 19524 rows and 6 columns.
}
\usage{
DrugBank_targets
}
\description{
DrugBank_targets
}
\keyword{datasets}
