% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugs_PO_FDA_biomarkers}
\alias{drugs_PO_FDA_biomarkers}
\title{drugs_PO_FDA_biomarkers}
\format{
An object of class \code{data.frame} with 107 rows and 10 columns.
}
\usage{
drugs_PO_FDA_biomarkers
}
\description{
drugs_PO_FDA_biomarkers
}
\keyword{datasets}
