% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FDA_approved_drugs}
\alias{FDA_approved_drugs}
\title{FDA_approved_drugs}
\format{
An object of class \code{character} of length 4346.
}
\usage{
FDA_approved_drugs
}
\description{
FDA_approved_drugs
}
\keyword{datasets}
