% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{filter_drugs_select_genes}
\alias{filter_drugs_select_genes}
\title{filter data frames based on the gene/protein being in keep_genes
keep only the unique drug names}
\usage{
filter_drugs_select_genes(drugs_genes_df, keep_genes)
}
\arguments{
\item{drugs_genes_df}{data.frame contains gene/protein names}

\item{keep_genes}{list of genes to filter or keep}
}
\value{
data.frame
}
\description{
filter data frames based on the gene/protein being in keep_genes
keep only the unique drug names
}
