% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{gene_names_standard}
\alias{gene_names_standard}
\title{function that inputs a list of genes and outputs a vector matching the names to the standardized names
for the genes that were matched up using DGI (and just returns the original names for the ones that were not)}
\usage{
gene_names_standard(genes)
}
\arguments{
\item{genes}{list of gene names to standardize}
}
\value{
list of gene names
}
\description{
function that inputs a list of genes and outputs a vector matching the names to the standardized names
for the genes that were matched up using DGI (and just returns the original names for the ones that were not)
}
