% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_cat_1_2}
\alias{get_cat_1_2}
\title{get Category 1,2 recommendations
if cat2 == "yes", then it's category 2, otherwise it's category 1
(difference is in using given cancer type, vs. all other cancer types)}
\usage{
get_cat_1_2(MP, cancer_type, drugs_PO_FDA_biomarkers, cat2 = "no")
}
\arguments{
\item{MP}{input molecular profile}

\item{cancer_type}{cancer type selection}

\item{drugs_PO_FDA_biomarkers}{drugs data frame}

\item{cat2}{}
}
\description{
get Category 1,2 recommendations
if cat2 == "yes", then it's category 2, otherwise it's category 1
(difference is in using given cancer type, vs. all other cancer types)
}
