% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_cat_3_4}
\alias{get_cat_3_4}
\title{get Category 3,4 recommendations
if cat4 == "yes", then it's category 4, otherwise it's category 3
(difference is in using given cancer type, vs. all other cancer types)}
\usage{
get_cat_3_4(
  MP,
  cancer_type,
  cat_drugs,
  list_paths,
  Onc_df,
  drugs_PO_FDA_biomarkers,
  drugs_PO_FDA_targets,
  drug_targets,
  FDA_approved_drugs,
  Type1,
  Type2,
  Type3 = NULL,
  cat4 = "no",
  subtype = "activation"
)
}
\arguments{
\item{MP}{input molecular profile}

\item{cancer_type}{cancer type selection}

\item{cat_drugs}{character string - whether only targeted cancer therapies, all FDA approved therapies, or all drugs in DrugBank}

\item{list_paths}{list of pathways}

\item{Onc_df}{data frame with oncogenes and cancer type}

\item{drugs_PO_FDA_biomarkers}{FDA-approved precision oncology drugs with listed biomarkers}

\item{drugs_PO_FDA_targets}{FDA-approved precision oncology drugs with targets}

\item{drug_targets}{data frame of targets for specific genes/proteins - right now, have DrugBank_targets}

\item{FDA_approved_drugs}{list of FDA-approved drugs)}

\item{Type1}{results of get_cat_1_2}

\item{Type2}{results of get_cat_1_2}

\item{Type3}{results of get_cat_3_4 (only use in cat4 == "yes)}

\item{cat4}{defaults to "no"}

\item{subtype}{choose subtype of connections, if have multiple ones in list_paths}
}
\value{
data frame
}
\description{
get Category 3,4 recommendations
if cat4 == "yes", then it's category 4, otherwise it's category 3
(difference is in using given cancer type, vs. all other cancer types)
}
