% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_connections_paths}
\alias{get_connections_paths}
\title{function that gets connections from one or more pathways which are within a list like the one of KEGG pathways}
\usage{
get_connections_paths(list_of_paths, IDs, subtype = NULL)
}
\arguments{
\item{list_of_paths}{is the list of pathways}

\item{IDs}{is a vector of IDs for the names of the pathways that should be pulled out}

\item{subtype}{may specify "activation" or just be NULL, in which case everything is included}
}
\value{
output is a data frame of connections
}
\description{
function that gets connections from one or more pathways which are within a list like the one of KEGG pathways
}
