% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_downstream_from_inputs}
\alias{get_downstream_from_inputs}
\title{function to get nodes downstream from given nodes
starts from the entire pathway - calls calculate_shortest_path_from_inputs and
get_downstream_nodes}
\usage{
get_downstream_from_inputs(cancer_path, MP, inputs_in_path)
}
\description{
function to get nodes downstream from given nodes
starts from the entire pathway - calls calculate_shortest_path_from_inputs and
get_downstream_nodes
}
