% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_downstream_nodes}
\alias{get_downstream_nodes}
\title{function to get nodes downstream from given nodes (which are the inputs)}
\usage{
get_downstream_nodes(inputs, shortest_paths_from_inputs)
}
\arguments{
\item{inputs}{is a character vector}

\item{shortest_paths_from_inputs}{is matrix generated by calculate_shortest_path}
}
\value{
output is a list of downstream nodes for each input node
}
\description{
function to get nodes downstream from given nodes (which are the inputs)
}
