% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{get_edges_cats_3_4}
\alias{get_edges_cats_3_4}
\title{get edges for categories 3 and 4}
\usage{
get_edges_cats_3_4(Cat3, Cat4 = NULL)
}
\arguments{
\item{Cat3}{reactive objects corresponding to categories 3}

\item{Cat4}{reactive objects corresponding to categories 4, Cat4 may be null}
}
\value{
edges
}
\description{
get edges for categories 3 and 4
}
