% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{graph_pathways_cats_3_4}
\alias{graph_pathways_cats_3_4}
\title{function that generates pathway plots for categories 3 and 4}
\usage{
graph_pathways_cats_3_4(edges, drugs, inputs)
}
\arguments{
\item{drugs}{character vectors}

\item{inputs}{character vectors}

\item{edge}{is the matrix of edges}
}
\value{
pathway plot
}
\description{
function that generates pathway plots for categories 3 and 4
}
