% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_cancer_paths_onc_long}
\alias{KEGG_cancer_paths_onc_long}
\title{KEGG_cancer_paths_onc_long}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
KEGG_cancer_paths_onc_long
}
\description{
KEGG_cancer_paths_onc_long
}
\keyword{datasets}
