% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_paths_KEGG}
\alias{list_paths_KEGG}
\title{list_paths_KEGG}
\format{
An object of class \code{list} of length 16.
}
\usage{
list_paths_KEGG
}
\description{
list_paths_KEGG
}
\keyword{datasets}
