% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{make_df_edges_targeted}
\alias{make_df_edges_targeted}
\title{function that creates the data frame specifying the edges from a data frame like drugs_PO_FDA_keep}
\usage{
make_df_edges_targeted(df_targeted, subtype = "sensitivity")
}
\arguments{
\item{df_targeted}{data.drame}

\item{subtype}{default to sensitivity}
}
\value{
data.frame
}
\description{
function that creates the data frame specifying the edges from a data frame like drugs_PO_FDA_keep
}
