% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{make_node_df}
\alias{make_node_df}
\title{function that makes data frame of nodes}
\usage{
make_node_df(assoc_df)
}
\arguments{
\item{assoc_df}{is the data frame of drug-gene/protein assocations}
}
\value{
data frame
}
\description{
function that makes data frame of nodes
}
