% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{MP_add_genes}
\alias{MP_add_genes}
\title{function that adds other genes to the molecular profiling (MP) data frame}
\usage{
MP_add_genes(MP_df, add_genes)
}
\arguments{
\item{MP_df}{data.frame containing molecular profile}

\item{add_genes}{list of genes to add to profile}
}
\value{
data frame
}
\description{
function that adds other genes to the molecular profiling (MP) data frame
}
