% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_prop2prop}
\alias{non_prop2prop}
\title{non_prop2prop}
\format{
An object of class \code{character} of length 52211.
}
\usage{
non_prop2prop
}
\description{
non_prop2prop
}
\keyword{datasets}
