% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{paste_shortest_path}
\alias{paste_shortest_path}
\title{function that pastes the shortest paths between some inputs and some outputs
(inputs_in_path are downstream of the oncogenes)}
\usage{
paste_shortest_path(inputs, inputs_in_path, outputs, igraph_obj)
}
\arguments{
\item{igraph_obj}{is created by converting a data frame of edges using graph.data.frame}

\item{input}{character vectors}

\item{output}{character vector}
}
\value{
a character vector with the paths marked with "->", separated by ";" if there are multiple paths
}
\description{
function that pastes the shortest paths between some inputs and some outputs
(inputs_in_path are downstream of the oncogenes)
}
