% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{pretty_drugs_mat}
\alias{pretty_drugs_mat}
\title{function that prettifies output of drugs_mat}
\usage{
pretty_drugs_mat(drugs_mat)
}
\arguments{
\item{drug_mat}{a drug matrix}
}
\value{
data frame
}
\description{
function that prettifies output of drugs_mat
}
