% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_non_prop}
\alias{prop_non_prop}
\title{prop_non_prop}
\format{
An object of class \code{character} of length 52211.
}
\usage{
prop_non_prop
}
\description{
prop_non_prop
}
\keyword{datasets}
