% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{remove_known_recs}
\alias{remove_known_recs}
\title{for category 3 and 4 recommendations, take out the drugs already recommended above
(i.e. recommended in categories 1 and 2 for category 3 and in categories 1, 2, or 3
for category 4)
if Type3_df == NULL, then this means just take out category 1 and 2 drugs + anything that has
(otherwise, also take out category 3 drugs)}
\usage{
remove_known_recs(
  drugs_biomarkers_targets,
  Type1_df,
  Type2_df,
  Type3_df = NULL
)
}
\arguments{
\item{drugs_biomarkers_targets}{is the initial input data frame, from which known recommendations will be removed}

\item{Type1_df}{category 1 drug data frame}

\item{Type2_df}{category 2 drug data frame}

\item{Type3_df}{category 3 drug data frame}
}
\value{
filtered data frame
}
\description{
for category 3 and 4 recommendations, take out the drugs already recommended above
(i.e. recommended in categories 1 and 2 for category 3 and in categories 1, 2, or 3
for category 4)
if Type3_df == NULL, then this means just take out category 1 and 2 drugs + anything that has
(otherwise, also take out category 3 drugs)
}
