% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-functions.R
\name{remove_WT}
\alias{remove_WT}
\title{function to remove wild type alterations from an MP data frame}
\usage{
remove_WT(MP_df, wild_type = c("wild type", "wildtype", "wild", "wt"))
}
\arguments{
\item{MP_df}{data frame with molecular alterations}

\item{wild_type}{ways in which the wild type can be coded in MP_df}
}
\value{
data frame
}
\description{
function to remove wild type alterations from an MP data frame
}
