% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdgnet-shiny.R
\name{router}
\alias{router}
\title{router for the application to extract parameters from the url}
\usage{
router(input, output, session = shiny::getDefaultReactiveDomain())
}
\description{
router for the application to extract parameters from the url
}
